Mobile&CyberPhysicalSystems/BlueFace_project/auth_factors/requirements.txt
certifi==2020.4.5.1
click==7.1.2
dlib==19.19.0
dnspython==1.16.0
face-recognition==1.3.0
face-recognition-models==0.3.0
ipaddress==1.0.23
numpy==1.18.3
paho-mqtt==1.5.0
Pillow==7.1.2
PyBluez==0.23
pymongo==3.10.1
pyobjc-core==6.2
pyobjc-framework-Cocoa==6.2
timeloop==1.0.2
wincertstore==0.2
